refactored db and crypto folders/exports
