+++ src/crypto/rateCommitmentHasher.ts
@@ -1,7 +1,8 @@
 import { poseidon2 } from 'poseidon-lite/poseidon2';
 
-function getRateCommitmentHash(identityCommitment: bigint, userMessageLimit: number | bigint) {
+export function getRateCommitmentHash(
+  identityCommitment: bigint,
+  userMessageLimit: number | bigint
+): bigint {
   return poseidon2([identityCommitment, userMessageLimit]);
 }
-
-export default getRateCommitmentHash;
