+++ src/data/messages.ts
@@ -1,11 +1,8 @@
-import { removeIdentityFromRoom } from './db';
-import { PrismaClient } from '@prisma/client';
+import { createMessageInRoom, findRoomWithMessageId, removeIdentityFromRoom } from './db/';
 import { MessageI, RoomI } from 'discreetly-interfaces';
 import { shamirRecovery, getIdentityCommitmentFromSecret } from '../crypto/shamirRecovery';
 import { RLNFullProof } from 'rlnjs';
-import verifyProof from '../crypto/verifier';
-
-const prisma = new PrismaClient();
+import { verifyProof } from '../crypto/';
 
 interface CollisionCheckResult {
   collision: boolean;
@@ -22,98 +19,46 @@ interface CollisionCheckResult {
  */
 
 async function checkRLNCollision(roomId: string, message: MessageI): Promise<CollisionCheckResult> {
-  return new Promise((res) => {
-    prisma.rooms
-      .findFirst({
-        where: { roomId },
-        include: {
-          epochs: {
-            where: { epoch: String(message.epoch) },
-            include: {
-              messages: {
-                where: { messageId: message.messageId }
-              }
-            }
-          }
-        }
-      })
-      .then((oldMessage) => {
-        if (!message.proof) {
-          throw new Error('Proof not provided');
-        }
+  const oldMessage: MessageI | null = await findRoomWithMessageId(roomId, message);
 
-        if (
-          !oldMessage ||
-          !oldMessage?.epochs[0]?.messages ||
-          !oldMessage?.epochs[0]?.messages[0] ||
-          !oldMessage?.epochs[0]?.messages[0]?.proof
-        ) {
-          console.debug('No collision', oldMessage);
-          res({ collision: false } as CollisionCheckResult);
-        } else {
-          const oldMessageProof = JSON.parse(
-            oldMessage.epochs[0].messages[0].proof
-          ) as RLNFullProof;
-          const oldMessagex2 = BigInt(oldMessageProof.snarkProof.publicSignals.x);
-          const oldMessagey2 = BigInt(oldMessageProof.snarkProof.publicSignals.y);
-
-          let proof: RLNFullProof;
-
-          if (typeof message.proof === 'string') {
-            proof = JSON.parse(message.proof) as RLNFullProof;
-          } else {
-            proof = message.proof;
-          }
-          const [x1, y1] = [
-            BigInt(proof.snarkProof.publicSignals.x),
-            BigInt(proof.snarkProof.publicSignals.y)
-          ];
-          const [x2, y2] = [oldMessagex2, oldMessagey2];
-
-          const secret = shamirRecovery(x1, x2, y1, y2);
-
-          res({
-            collision: true,
-            secret,
-            oldMessage: oldMessage.epochs[0].messages[0] as unknown as MessageI
-          } as CollisionCheckResult);
-        }
-      })
-      .catch((err) => console.error(err));
-  });
-}
-
-/**
- * Adds a message to a room.
- * @param {string} roomId - The ID of the room to add the message to.
- * @param {MessageI} message - The message to add to the room.
- * @returns {Promise<unknown>} - A promise that resolves when the message has been added to the room.
- */
-
-function addMessageToRoom(roomId: string, message: MessageI): Promise<unknown> {
-  if (!message.epoch) {
-    throw new Error('Epoch not provided');
+  if (!message.proof) {
+    throw new Error('Proof not provided');
   }
-  return prisma.rooms.update({
-    where: {
-      roomId: roomId
-    },
-    data: {
-      epochs: {
-        create: {
-          epoch: String(message.epoch),
-          messages: {
-            create: {
-              message: message.message ? String(message.message) : '',
-              messageId: message.messageId ? message.messageId.toString() : '',
-              proof: JSON.stringify(message.proof),
-              roomId: roomId
-            }
-          }
-        }
-      }
+
+  if (!oldMessage?.proof) {
+    console.debug('No collision', oldMessage);
+    return { collision: false } as CollisionCheckResult;
+  } else {
+    let oldMessageProof: RLNFullProof;
+    if (typeof oldMessage.proof === 'string') {
+      oldMessageProof = JSON.parse(oldMessage.proof) as RLNFullProof;
+    } else {
+      oldMessageProof = oldMessage.proof;
     }
-  });
+    const oldMessagex2 = BigInt(oldMessageProof.snarkProof.publicSignals.x);
+    const oldMessagey2 = BigInt(oldMessageProof.snarkProof.publicSignals.y);
+
+    let proof: RLNFullProof;
+
+    if (typeof message.proof === 'string') {
+      proof = JSON.parse(message.proof) as RLNFullProof;
+    } else {
+      proof = message.proof;
+    }
+    const [x1, y1] = [
+      BigInt(proof.snarkProof.publicSignals.x),
+      BigInt(proof.snarkProof.publicSignals.y)
+    ];
+    const [x2, y2] = [oldMessagex2, oldMessagey2];
+
+    const secret = shamirRecovery(x1, x2, y1, y2);
+
+    return {
+      collision: true,
+      secret,
+      oldMessage: oldMessage
+    } as CollisionCheckResult;
+  }
 }
 
 export interface validateMessageResult {
@@ -130,7 +75,7 @@ async function handleCollision(
   const roomId = room.roomId.toString();
   if (!collisionResult.collision) {
     try {
-      await addMessageToRoom(roomId, message);
+      await createMessageInRoom(roomId, message);
       console.debug(
         `Message added to room: ${
           typeof message.message === 'string'
@@ -167,7 +112,7 @@ export async function validateMessage(
   message: MessageI
 ): Promise<validateMessageResult> {
   const roomId = room.roomId.toString();
-  const validProof = await verifyProof(message, room);
+  const validProof: boolean = await verifyProof(room, message);
   if (validProof) {
     const collisionResult = await checkRLNCollision(roomId, message);
     const result = await handleCollision(room, message, collisionResult);
