+++ src/types/index.ts
@@ -1,1 +1,16 @@
+import { RoomI } from 'discreetly-interfaces';
+
+export interface CodeStatus {
+  claimed: boolean;
+  roomIds: string[];
+}
+
+export interface RoomsFromClaimCode {
+  roomIds: string[];
+}
+
+export interface RoomWithSecretsI extends RoomI {
+  bandadaAPIKey: string;
+}
+
 export type userCountI = Record<string, number>;
